requirements.txt
astro-sdk-python
apache-airflow-providers-cncf-kubernetes
pendulum
astronomer-providers